requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-model-utils
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
django-cors-headers 
psycopg2
PyJWT